+++ docs/tutorial/events.js
@@ -58,7 +58,7 @@ Promise.all([dataLoaded, mdChartLoaded]).then((results) => {
         },
         (chart) => {
           chart.on('click', clickHandler)
-          chart.forceUpdate(true)
+          chart.feature.rendering.update()
           return chart
         }
       ]
@@ -81,7 +81,7 @@ Promise.all([dataLoaded, mdChartLoaded]).then((results) => {
         },
         (chart) => {
           chart.on('plot-axis-label-draw', labelDrawHandler)
-          chart.forceUpdate(true)
+          chart.feature.rendering.update()
           return chart
         }
       ]
@@ -104,7 +104,7 @@ Promise.all([dataLoaded, mdChartLoaded]).then((results) => {
         },
         (chart) => {
           chart.on('logo-draw', logoDrawHandler)
-          chart.forceUpdate(true)
+          chart.feature.rendering.update()
           return chart
         }
       ]
@@ -128,7 +128,7 @@ Promise.all([dataLoaded, mdChartLoaded]).then((results) => {
         (chart) => {
           const registerHandler = () => {
             chart.on('background-draw', backgroundImageHandler)
-            chart.forceUpdate(true)
+            chart.feature.rendering.update()
           }
           if (!image.complete) image.onload = registerHandler
           else registerHandler()