+++ src/apps/weblib/ts-api/plugins/cssproperties.ts
@@ -4,12 +4,14 @@ import { Snapshot } from '../module/cchart.js'
 import Vizzu from '../vizzu.js'
 
 import { getCSSCustomPropsForElement, propsToObject } from './cssutils.js'
+import { HtmlCanvasApi } from '../htmlcanvas.js'
 
 export class CSSProperties implements Plugin {
   private _chart?: Vizzu
 
   meta = {
-    name: 'cssProperties'
+    name: 'cssProperties',
+    depends: ['htmlCanvas']
   }
 
   api = {
@@ -25,7 +27,10 @@ export class CSSProperties implements Plugin {
   get hooks(): PluginHooks {
     return {
       prepareAnimation: (ctx: PrepareAnimationContext, next: () => void): void => {
-        const props = getCSSCustomPropsForElement(this._chart!.getCanvasElement(), this.api.prefix)
+        const props = getCSSCustomPropsForElement(
+          (this._chart!.feature['htmlCanvas'] as HtmlCanvasApi).element(),
+          this.api.prefix
+        )
         if (Array.isArray(ctx.target))
           ctx.target.forEach(({ target }) => {
             if (!(target instanceof Snapshot)) {